+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/primitive/repair/SupplyOfflineSelector.java
@@ -7,16 +7,14 @@ import net.splitcells.gel.constraint.GroupId;
 import net.splitcells.gel.data.table.Line;
 import net.splitcells.gel.solution.optimization.OfflineOptimization;
 
-import java.util.Optional;
 import java.util.function.BiFunction;
 
 import static net.splitcells.dem.data.set.Sets.setOfUniques;
 import static net.splitcells.dem.data.set.list.Lists.listWithValuesOf;
 import static net.splitcells.dem.utils.random.RandomnessSource.randomness;
-import static net.splitcells.gel.solution.optimization.primitive.SupplySelection.supplySelection;
 
 @FunctionalInterface
-public interface SupplySelector extends BiFunction<Map<GroupId, Set<Line>>, List<Line>, OfflineOptimization> {
+public interface SupplyOfflineSelector extends BiFunction<Map<GroupId, Set<Line>>, List<Line>, OfflineOptimization> {
 
     OfflineOptimization apply(Map<GroupId, Set<Line>> freeDemandGroups, List<Line> freeSupplies);
 }