+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/primitive/repair/SupplyOfflineSelectors.java
@@ -2,7 +2,6 @@ package net.splitcells.gel.solution.optimization.primitive.repair;
 
 import net.splitcells.dem.data.set.Set;
 import net.splitcells.dem.data.set.Sets;
-import net.splitcells.dem.environment.config.StaticFlags;
 import net.splitcells.gel.data.table.Line;
 import net.splitcells.gel.solution.optimization.OptimizationEvent;
 import net.splitcells.gel.solution.optimization.primitive.SupplySelection;
@@ -19,12 +18,12 @@ import static net.splitcells.gel.solution.optimization.StepType.ADDITION;
 import static net.splitcells.gel.solution.optimization.primitive.SupplySelection.supplySelection;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class SupplySelectors {
-    private SupplySelectors() {
+public class SupplyOfflineSelectors {
+    private SupplyOfflineSelectors() {
         throw constructorIllegal();
     }
 
-    public static SupplySelector indexBasedRepairer
+    public static SupplyOfflineSelector indexBasedRepairer
             (BiFunction<Integer, Integer, Optional<SupplySelection>> indexSelector) {
         return (freeDemandGroups, freedSupplies) -> solution -> {
             final Set<OptimizationEvent> repairs = setOfUniques();
@@ -60,7 +59,7 @@ public class SupplySelectors {
         };
     }
 
-    public static SupplySelector supplySelector() {
+    public static SupplyOfflineSelector supplySelector() {
         final var randomness = randomness();
         return indexBasedRepairer((freeSupplyCount, supplyFreedCount) -> {
             if (freeSupplyCount.floatValue() + supplyFreedCount.floatValue() <= 0) {