+++ projects/net.splitcells.gel.core/CHANGELOG.md
@@ -2,6 +2,9 @@
 ## [Unreleased]
 ### Major Changes
 * **2022-03-31** **\#8**:
-  Rename `ConstraintGroupBasedRepair` to `ConstraintGroupBasedOfflineRepair`,
-  so it will be easy to distinguish this optimizer with the planned online
-  version.
+  Rename optimizer components,
+  so it will be easy to distinguish the offline repair from the planned online repair:
+  1. Rename `ConstraintGroupBasedRepair` to `ConstraintGroupBasedOfflineRepair`.
+  2. Rename `SupplySelector` to `SupplyOfflineSelector`.
+  3. Rename `SupplySelectors` to `SupplyOfflineSelectors`.
+